pyproject.toml
[project]
name = "payouts-service"
version = "0.1.0"
description = "Payouts REST service"
requires-python = ">=3.12"

dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.15",
    "celery>=5.4",
    "redis>=5.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "gunicorn>=23.0.0,<24.0.0",
    "drf-spectacular>=0.29.0,<0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "black>=24.2",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
